requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
ansicon==1.89.0
async-timeout==4.0.2
attrs==22.2.0
blessed==1.20.0
charset-normalizer==3.1.0
codefind==0.1.3
discord-py-interactions==5.0.0
discord-typings==0.5.1
emoji==2.2.0
frozenlist==1.3.3
idna==3.4
iso8601==1.1.0
jinxed==1.2.0
jurigged==0.5.5
multidict==6.0.4
ovld==0.3.2
six==1.16.0
tomli==2.0.1
twitchio==2.6.0
typing_extensions==4.5.0
watchdog==3.0.0
wcwidth==0.2.6
yarl==1.8.2
